==> code/01-Data_Collection_and_EDA.ipynb <==
## PushShift
#### Pulling Data for r/MMA and r/UFC
---


import requests
import calendar
import time
import pandas as pd
from bs4 import BeautifulSoup
import re

url = 'https://api.pushshift.io/reddit/search/submission'

# doing a preliminary pull to see what I need to filter out in future pulls
# and in helper functions
params = {
    'subreddit': 'UFC, MMA',
    'limit': 10,
}
res = requests.get(url, params)
print(res.status_code)
df = pd.DataFrame(res.json()['data'])


#checking out all the data provided to narrow down the filter param
pd.set_option('display.max_rows', None)

df.tail().T

# pulling different features of interest through each call

def get_subreddit_mma_ufc_data(until):
    '''
    function that returns a dataframe with 1000 entries using
    api pushshift params. The UNTIL parameter is used to help with pulling non-overlapping
    entries.
    '''
    
    params = {
        'subreddit': 'UFC, MMA',
        'limit': 1000,
        'filter': 'subreddit, id, selftext, title, media, media_embed, num_comments, created_utc',
        'until': until
    }
    res = requests.get(url, params)
    if res.status_code == 200:
        df = pd.DataFrame(res.json()['data'])
        return df

#this UNTIL argument was the time (in utc) when it was first used
df01 = get_subreddit_mma_ufc_data(1673205797)
df01.tail(1)

#the UNTIL arguments were found after calling these separately and looking at the
# .tail(1) of the dfs and selecting the last created_utc.
df02 = get_subreddit_mma_ufc_data(1672789239)
df03 = get_subreddit_mma_ufc_data(1672318448)
df04 = get_subreddit_mma_ufc_data(1671808988)
df05 = get_subreddit_mma_ufc_data(1671484146)
df06 = get_subreddit_mma_ufc_data(1671116165)
df07 = get_subreddit_mma_ufc_data(1670800057)
df08 = get_subreddit_mma_ufc_data(1670737008)

#combining into a comprehensive df

df_list = [df01, df02, df03, df04, df05, df06, df07, df08]

df_total = pd.concat(df_list, ignore_index = True)

#looking at the proportions of data for each subreddit
df_total['subreddit'].value_counts(normalize = True)

#trying to even out the proportions. Writing a similar function to the previous,
# but only for data from r/MMA

def just_mma_data(until):
    '''
    To get more even number of data, I am searching for only MMA data
    '''
    params2 = {
        'subreddit': 'MMA',
        'limit': 1000,
        'filter': 'subreddit, id, selftext, title, media, media_embed, num_comments, created_utc',
        'until': until
    }
    res = requests.get(url, params2)
    if res.status_code == 200:
        df = pd.DataFrame(res.json()['data'])
        return df

df09 = just_mma_data(1670598695)
df10 = just_mma_data(1669478929)
df11 = just_mma_data(1668574869)
df = pd.concat([df_total, df09, df10, df11], ignore_index = True)

# ensuring similar counts from each subreddit
df['subreddit'].value_counts(normalize = True)

#checking for reposts
    #therfore only checking title duplicates
    
df['title'].duplicated().sum()

df.drop_duplicates(subset = 'title', inplace = True)

df.shape

df.reset_index(inplace = True, drop = True)

# Trying to figure out if any of the retreived posts are ads or promotions
    # by checking to see if certain words are present

words = ['[Ss]ponsor', '\s[Aa]d\s', '[Pp]romoted']
counters = []
for word in words:
    counter = 0
    for i in df['title']:
        if len(re.findall(word, i)) != 0:
            counter +=1
    counters.append(counter)
pd.DataFrame(data = counters, index = words)

# Making a quick df to be able to see the indexes for my main df

words = ['[Ss]ponsor', '\s[Aa]d\s', '[Pp]romoted']
indexes = []
for word in words:
    index = []
    for i in range(0, len(df['title'])-1):
        if len(re.findall(word, df['title'][i])) != 0:
            index.append(i)
    indexes.append(index)
ad_eda = pd.DataFrame(data = indexes, index = words).T
ad_eda.fillna(0, inplace = True)
ad_eda.astype(int)

# Looking at various entries to see if the wordsearches were indeed ads

for i in ad_eda.columns:
    for j in range(0, 6, 3):
        if ad_eda[i][j] != 0:
            print(df['title'][(ad_eda[i][j])]+'\n')

> It doesnt look like these are ads or promotions, which is excellent

df[df['selftext'] == ''].count()

Approximately 75% of the observations have nothing in the 'selftext' colunmn

# Interesting that subreddit MMA has many more [removed] selftexts. Which could be used in
#    a model. So instead of replacing with an empty string, I will keep [removed] in.
df['selftext'][df['selftext'] == '[removed]'].groupby(df['subreddit']).count()

df['selftext'] = df['selftext'].map(lambda x: 1 if x =='[removed]' else 0)
df['selftext'].value_counts(normalize=True)

######
---
######
#### Checking for UFC-Specific words as well as the other MMA league, Bellator
######

words = ['[Dd]ana', '[Ww]hite', '[Uu][Ff][Cc]', '[Uu]ltimate', '[Bb]ellator', '[Mm]c[Gg]regor', '[Aa]desanya', '[Pp]addy']
indexes = []
for word in words:
    index = []
    for i in range(0, len(df['title'])-1):
        if len(re.findall(word, df['title'][i])) != 0:
            index.append(i)
    indexes.append(index)
ufc_eda = pd.DataFrame(data = indexes, index = words).T
ufc_eda = ufc_eda.fillna(0)
ufc_eda = ufc_eda.astype(int)


ufc_eda[ufc_eda != 0].count()

for i in ufc_eda.columns:
    for j in range(0, 1993, 150):
        if ufc_eda[i][j] != 0:
            print(df['title'][(ufc_eda[i][j])]+'|'+df['subreddit'][(ufc_eda[i][j])]+'\n')

words = ['[Dd]ana', '[Ww]hite', '[Uu][Ff][Cc]', '[Uu]ltimate', '[Bb]ellator', '[Mm]c[Gg]regor', '[Aa]desanya', '[Pp]addy']
indexes = []
for word in words:
    index = []
    for i in range(0, len(df['title'])-1):
        if len(re.findall(word, df['title'][i])) != 0:
            if df['subreddit'][i] == 'MMA':
                index.append(i)
    indexes.append(index)
mma_eda = pd.DataFrame(data = indexes, index = words).T
mma_eda = mma_eda.fillna(0)
mma_eda = mma_eda.astype(int)

mma_eda[mma_eda != 0].count()

df['subreddit'].value_counts()

1126/5057

22.3% of ALL r/mma titles includes the word 'ufc'

---

df['subreddit'].value_counts(normalize = True)

###
#### This represents the Null Model Accuracy, which is 50.3%
###

##### In 22.27% of all MMA subreddit posts, 'UFC' is present

#Making the word count as a separate column
df['word_count'] = df['title'].map(lambda i: len(i.split(' ')))

# Binarizing subreddits
df['subreddit'] = df['subreddit'].map({'ufc':1, 'MMA': 0})

#exporting to csv for further preprocessing and modeling

df.to_csv('../data/ufc_mma_submissions.csv', index = False)
