==> code/02-Preprocessing_and_Initial_Modeling.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

from sklearn.model_selection import train_test_split, cross_val_score, GridSearchCV, cross_val_predict, RandomizedSearchCV
from sklearn.pipeline import Pipeline
from sklearn.naive_bayes import MultinomialNB
from sklearn.linear_model import LogisticRegression
from sklearn.metrics import confusion_matrix, plot_confusion_matrix
from sklearn.ensemble import RandomForestClassifier, StackingClassifier, AdaBoostClassifier
from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
from sklearn.tree import DecisionTreeClassifier
from sklearn.feature_extraction import text
import dataframe_image as dfi


df = pd.read_csv('../data/ufc_mma_submissions.csv')

X = df['title']
y = df['subreddit']

X_train, X_test, y_train, y_test = train_test_split(X, y, stratify = y, random_state=42)

pipe1 = Pipeline([
    ('cvec', CountVectorizer()),
    ('nb', MultinomialNB())
])

pipe1_params = {
    'cvec__max_features' : [5000, 6000, 7000, 8000, 9000, 10000, 11000, 12000],
    'cvec__min_df': [2, 3],
    'cvec__stop_words': [None, 'english'],
    'cvec__max_df' : [0.9, 0.95, 0.98],
    'cvec__ngram_range': [(1,1), (1,2)]
}

gs_pipe1 = GridSearchCV(pipe1,
                        param_grid=pipe1_params,
                        cv = 5,
                        n_jobs = -1)

gs_pipe1.fit(X_train, y_train)

print(f'Best Score: {gs_pipe1.best_score_}')
print(f'Best Params: {gs_pipe1.best_params_}')

print(f'Training Score: {gs_pipe1.score(X_train, y_train)}')
print(f'Testing Score: {gs_pipe1.score(X_test, y_test)}')

---

X = df['title']
y = df['subreddit']

X_train, X_test, y_train, y_test = train_test_split(X, y, stratify = y, random_state=42)

pipe2 = Pipeline([
    ('tvec', TfidfVectorizer()),
    ('nb', MultinomialNB())
])

pipe2_params = {
    'tvec__max_features' : [5000, 6000, 7000, 8000, 9000, 10000, 11000, 12000],
    'tvec__stop_words': [None, 'english'],
    'tvec__ngram_range': [(1,1), (1,2)],
    'tvec__max_df': [0.9, 0.95, 0.98],
    'tvec__min_df': [2, 3]
}

gs_pipe2 = GridSearchCV(pipe2,
                        param_grid=pipe2_params,
                        cv = 5,
                        n_jobs = -1)

gs_pipe2.fit(X_train, y_train)

print(f'Best Score: {gs_pipe2.best_score_}')
print(f'Best Params: {gs_pipe2.best_params_}')

print(f'Training Score: {gs_pipe2.score(X_train, y_train)}')
print(f'Testing Score: {gs_pipe2.score(X_test, y_test)}')

---

## Top Words Before Adding Stop Words

cvec = CountVectorizer(max_features=10000,
                             max_df = 0.9,
                             min_df = 2,
                             ngram_range=(1,2),
)
vectorized = cvec.fit_transform(X)
vectorized = pd.DataFrame(vectorized.todense(), columns = cvec.get_feature_names_out())
top_ufc = vectorized[df['subreddit'] == 1].sum().sort_values(ascending = False)[:20]
top_mma = vectorized[df['subreddit'] == 0].sum().sort_values(ascending = False)[:20]

count_index = []
ufc_count = []
mma_count = []
for i in top_ufc.index:
    if i in top_mma.index:
        count_index.append(i)
        ufc_count.append(top_ufc[i])
        mma_count.append(top_mma[i])

top_words = pd.DataFrame(index = count_index)
top_words['ufc'] = ufc_count
top_words['mma'] = mma_count

dfi.export(top_words, "../images/topwords_no_stop_words.png")
top_words

---

### Custom Stop Words

my_words_list= (['ufc', 'dana', 'white', 'ultimate', 'u.f.c.', 'islam', 'makhachev', 'moreno', 'edwards', 'usman', 'ngannou', 'adesanya',
                  'pantoja', 'kara', 'kai', 'oliveira', 'pereira', 'sterling', 'royval', 'nicolau', 'perez', 'albazi', 'schnell', 'omalley', 'yan',
                  'dvalishvili', 'vera', 'sandhagen', 'font', 'cruz', 'holloway', 'volkanovski', 'figueiredo', 'deiveson', 'aljamain', 'rodriguez',
                  'ortega', 'allen', 'emmett', 'chan', 'sung', 'jung', 'kattar', 'giga', 'chikadze', 'poirier', 'jones', 'elliott', 'dvorak', 'molina', 'mokaev',
                  'ulanbekov', 'yanez', 'gutierrez', 'nurmagomedov', 'simon', 'munhoz', 'shore', 'topuria', 'evloev', 'mitchell', 'yusuff', 'iga', 'barboza',
                  'caceres', 'burns', 'neal', 'luque', 'fiziev', 'gamrot', 'anjos', 'tsarukyan', 'turner', 'hooker', 'ismagulov', 'gaethje', 'magny', 'whittaker',
                  'vettori', 'strickland', 'costa', 'hermansson', 'covington', 'muniz', 'imavov', 'bachowicz', 'rakic', 'cannonier', 'dolidze', 'brunson', 'oezdemir',
                  'spann', 'walker', 'nunes', 'weili', 'shevchenko', 'pena', 'blaydes', 'tuivasa', 'aspinall', 'andrade', 'santos', 'daukaus', 'tybura', 'lewis', 'holm',
                  'vieira', 'jandiroba', 'maia', 'grasso', 'chookagian', 'murphy', 'fiorot', 'lemos', 'namajunas', 'esparza', 'jandiroba', 'blanchfield', 'barber',
                  'calvillo', 'ribas', 'viana', 'ducote', 'pinheiro', 'xiaonan', 'yan', 'abdurakhimov', 'spivac', 'shamil', 'ketlen', 'pennington', 'miesha', 'kunitskaya',
                  'rosa', 'avila', 'lansberg', 'paddy', 'silva', 'cormier', 'diaz', 'miocic', 'lesnar', 'penn', 'liddell', 'pierre', 'rousey', 'khabib', 'conor', 'mcgregor',
                  'frevola', 'dillashaw', 'pimblett', 'helwani', 'blachowicz','arlovski', 'donatello', 'dec', 'december', 'jan', 'feb', 'selftext', 'says', 'did', 'does',
                  'guy', 'guys', 'know', 'fc', 'vs', 'https', 'khamzat', '2022', '2023', '219', '281', '282', '283', '284', '285', 'going', 'man', 'got', 'anne', 'didnt', 
                  'ufc281', 'ankalaev', 'zhang', 'israel', 'johnson', 'dustin', 'krause', 'chandler', 'jiri', 'cejudo', 'march', 'februrary', 'gordon', 'ilia', 'florian',
                  'makachov', 'beneil', 'dariush', 'jared'])
stop_words_list = text.ENGLISH_STOP_WORDS.union(my_words_list)

# Got help with this combination from https://stackoverflow.com/questions/26826002/adding-words-to-stop-words-list-in-tfidfvectorizer-in-sklearn
# UFC fighters taken from https://www.ufc.com/rankings

X = df['title']
y = df['subreddit']

X_train, X_test, y_train, y_test = train_test_split(X, y, stratify = y, random_state=42)

pipe4 = Pipeline([
    ('cvec', CountVectorizer(stop_words=stop_words_list)),
    ('nb', MultinomialNB())
])

pipe4_params = {
    'cvec__max_features': [3000, 4000, 5000],
    'cvec__ngram_range': [(1,1), (1,2)],
    'cvec__max_df': [0.01, 0.05, 0.1],
    'cvec__min_df': [2, 3],
}

gs_pipe4 = GridSearchCV(pipe4,
                        param_grid=pipe4_params,
                        cv = 5)

gs_pipe4.fit(X_train, y_train)

print(f'Best Score: {gs_pipe4.best_score_}')
print(f'Best Params: {gs_pipe4.best_params_}')
print(f'Training Score: {gs_pipe4.score(X_train, y_train)}')
print(f'Testing Score: {gs_pipe4.score(X_test, y_test)}')

pipe3 = Pipeline([
    ('tvec', TfidfVectorizer(stop_words=stop_words_list)),
    ('nb', MultinomialNB())
])

pipe3_params = {
    'tvec__max_features': [2000, 4000, 6000, 8000, 10000],
    'tvec__ngram_range': [(1,1), (1,2)],
    'tvec__max_df': [0.8, 0.9, 0.95, 0.98, 1.0],
    'tvec__min_df': [2, 3, 4],
}

gs_pipe3 = GridSearchCV(pipe3,
                        param_grid=pipe3_params,
                        cv = 5)

gs_pipe3.fit(X_train, y_train)

print(f'Best Score: {gs_pipe3.best_score_}')
print(f'Best Params: {gs_pipe3.best_params_}')

print(f'Training Score: {gs_pipe3.score(X_train, y_train)}')
print(f'Testing Score: {gs_pipe3.score(X_test, y_test)}')

---

## Stacked Models

lvl1_est_1 = [
    ('nb', MultinomialNB()),
    ('rf', RandomForestClassifier()),
    ('ada', AdaBoostClassifier())
]

lvl1_est_2 = [
    ('nb', MultinomialNB()),
    ('rf', RandomForestClassifier()),
    ('logr', LogisticRegression(max_iter=1000))
]

lvl1_est_3 = [
    ('nb', MultinomialNB()),
    ('logr', LogisticRegression(max_iter=1000)),
    ('ada', AdaBoostClassifier())
]

lvl1_est_4 = [
    ('logr', LogisticRegression(max_iter=1000)),
    ('rf', RandomForestClassifier()),
    ('ada', AdaBoostClassifier())
]

stacked_1 = StackingClassifier(estimators=lvl1_est_1, 
                               final_estimator=LogisticRegression(),
                               n_jobs = -1)

pipe_tvec_1 = Pipeline([
    ('tvec', TfidfVectorizer(stop_words=stop_words_list)),
    ('s1', stacked_1)
])

print(cross_val_score(pipe_tvec_1, X_train, y_train).mean())
pipe_tvec_1.fit(X_train, y_train)
print(f'Training Score: {pipe_tvec_1.score(X_train, y_train)}')
print(f'Testing Score: {pipe_tvec_1.score(X_test, y_test)}')

stacked_2 = StackingClassifier(estimators=lvl1_est_2, 
                               final_estimator=LogisticRegression(),
                               n_jobs = -1)
pipe_tvec_2 = Pipeline([
    ('tvec', TfidfVectorizer(stop_words=stop_words_list)),
    ('s2', stacked_2)
])

print(cross_val_score(pipe_tvec_2, X_train, y_train).mean())
pipe_tvec_2.fit(X_train, y_train)
print(f'Training Score: {pipe_tvec_2.score(X_train, y_train)}')
print(f'Testing Score: {pipe_tvec_2.score(X_test, y_test)}')

stacked_3 = StackingClassifier(estimators=lvl1_est_3, 
                               final_estimator=LogisticRegression(),
                               n_jobs = -1)
pipe_tvec_3 = Pipeline([
    ('tvec', TfidfVectorizer(stop_words=stop_words_list)),
    ('s3', stacked_3)
])

print(cross_val_score(pipe_tvec_3, X_train, y_train).mean())
pipe_tvec_3.fit(X_train, y_train)
print(f'Training Score: {pipe_tvec_3.score(X_train, y_train)}')
print(f'Testing Score: {pipe_tvec_3.score(X_test, y_test)}')

stacked_4 = StackingClassifier(estimators=lvl1_est_4, 
                               final_estimator=LogisticRegression(),
                               n_jobs = -1)
pipe_tvec_4 = Pipeline([
    ('tvec', TfidfVectorizer(stop_words=stop_words_list)),
    ('s4', stacked_4)
])

print(cross_val_score(pipe_tvec_4, X_train, y_train).mean())
pipe_tvec_4.fit(X_train, y_train)
print(f'Training Score: {pipe_tvec_4.score(X_train, y_train)}')
print(f'Testing Score: {pipe_tvec_4.score(X_test, y_test)}')

stacked_1 = StackingClassifier(estimators=lvl1_est_1, 
                               final_estimator=LogisticRegression(),
                               n_jobs = -1)
pipe_cvec_1 = Pipeline([
    ('cvec', CountVectorizer(stop_words=stop_words_list)),
    ('s1', stacked_1)
])

print(cross_val_score(pipe_cvec_1, X_train, y_train).mean())
pipe_cvec_1.fit(X_train, y_train)
print(f'Training Score: {pipe_cvec_1.score(X_train, y_train)}')
print(f'Testing Score: {pipe_cvec_1.score(X_test, y_test)}')

# BEST MODEL THUS FAR
stacked_2 = StackingClassifier(estimators=lvl1_est_2, 
                               final_estimator=LogisticRegression(),
                               n_jobs = -1)
pipe_cvec_2 = Pipeline([
    ('cvec', CountVectorizer(stop_words=stop_words_list)),
    ('s2', stacked_2)
])

print(cross_val_score(pipe_cvec_2, X_train, y_train).mean())
pipe_cvec_2.fit(X_train, y_train)
print(f'Training Score: {pipe_cvec_2.score(X_train, y_train)}')
print(f'Testing Score: {pipe_cvec_2.score(X_test, y_test)}')

stacked_3 = StackingClassifier(estimators=lvl1_est_3, 
                               final_estimator=LogisticRegression(),
                               n_jobs = -1)
pipe_cvec_3 = Pipeline([
    ('cvec', CountVectorizer(stop_words=stop_words_list)),
    ('s3', stacked_3)
])

print(cross_val_score(pipe_cvec_3, X_train, y_train).mean())
pipe_cvec_3.fit(X_train, y_train)
print(f'Training Score: {pipe_cvec_3.score(X_train, y_train)}')
print(f'Testing Score: {pipe_cvec_3.score(X_test, y_test)}')

stacked_4 = StackingClassifier(estimators=lvl1_est_4, 
                               final_estimator=LogisticRegression(),
                               n_jobs = -1)
pipe_cvec_4 = Pipeline([
    ('cvec', CountVectorizer(stop_words=stop_words_list)),
    ('s4', stacked_4)
])

print(cross_val_score(pipe_cvec_4, X_train, y_train).mean())
pipe_cvec_4.fit(X_train, y_train)
print(f'Training Score: {pipe_cvec_4.score(X_train, y_train)}')
print(f'Testing Score: {pipe_cvec_4.score(X_test, y_test)}')

#### The best model so far:
> the Stacked Model with CountVectorizer as the transformer, the first level estimators based on a Multinomial NB, a Random Forest Classifier, and an Logistic Classifier, and the final estimator being another Logistic Regression Model.

---
## Gridsearching Through the First Level Estimators for Best Parameters

### Starting With RandomSearch

Going through each individual estimator in the first level of my stacking model to find the best hyperparameters. This will hopefully mean my final stacking model will be even better.

#Starting with MultinomialNB
pipe1 = Pipeline([
    ('cvec', CountVectorizer(stop_words=stop_words_list)),
    ('nb', MultinomialNB())
])

pipe1_params = {
    'cvec__max_features' : range(500, 10001, 100),
    'cvec__min_df': range(1, 16),
    'cvec__max_df' : [0.9, 0.95, 0.98, 1],
    'cvec__ngram_range': [(1,1), (1,2), (2,2), (1,3)],
    'nb__alpha': [0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 1.0]
}

rs_pipe1 = RandomizedSearchCV(pipe1,
                        param_distributions=pipe1_params,
                        cv = 5)

rs_pipe1.fit(X_train, y_train)

print(f'Best Score: {rs_pipe1.best_score_}')
print(f'Best Params: {rs_pipe1.best_params_}')
print(f'Training Score: {rs_pipe1.score(X_train, y_train)}')
print(f'Testing Score: {rs_pipe1.score(X_test, y_test)}')

pipe1 = Pipeline([
    ('cvec', CountVectorizer(stop_words=stop_words_list)),
    ('nb', MultinomialNB())
])

pipe1_params = {
    'cvec__max_features' : [3000, 4000, 5000],
    'cvec__min_df': [2],
    'cvec__max_df' : [0.01, 0.05, 0.1, 0.15],
    'cvec__ngram_range': [(1,1)],
    'nb__alpha': [0.5, 0.6, 0.7]
}

gs_pipe1 = GridSearchCV(pipe1,
                        param_grid=pipe1_params,
                        cv = 5,
                        n_jobs= -1)

gs_pipe1.fit(X_train, y_train)

print(f'Best Score: {gs_pipe1.best_score_}')
print(f'Best Params: {gs_pipe1.best_params_}')
print(f'Training Score: {gs_pipe1.score(X_train, y_train)}')
print(f'Testing Score: {gs_pipe1.score(X_test, y_test)}')

pipe2 = Pipeline([
    ('cvec', CountVectorizer(stop_words=stop_words_list)),
    ('rf', RandomForestClassifier())
])

pipe2_params = {
    'cvec__max_features' : [8000, 8200, 8400],
    'cvec__min_df': [3, 4, 5],
    'cvec__max_df' : [0.9, 0.95, 0.98],
    'cvec__ngram_range': [(1,1), (1,2)],
    'rf__n_estimators': range(50, 10001, 50),
    'rf__max_depth': range(2, 101, 2),
    'rf__min_samples_leaf': range(1, 21, 1),
    'rf__min_samples_split': range(2, 101, 2)
}

rs_pipe2 = RandomizedSearchCV(pipe2,
                              param_distributions=pipe2_params,
                              cv = 5,
                              n_jobs = -1)

rs_pipe2.fit(X_train, y_train)

print(f'Best Score: {rs_pipe2.best_score_}')
print(f'Best Params: {rs_pipe2.best_params_}')
print(f'Training Score: {rs_pipe2.score(X_train, y_train)}')
print(f'Testing Score: {rs_pipe2.score(X_test, y_test)}')

pipe2 = Pipeline([
    ('cvec', CountVectorizer(stop_words=stop_words_list)),
    ('rf', RandomForestClassifier(n_jobs = -1))
])

pipe2_params = {
    'cvec__max_features' : [8000, 8200, 8400],
    'cvec__min_df': [3, 4, 5],
    'cvec__max_df' : [0.9, 0.95, 0.98],
    'cvec__ngram_range': [(1,1)],
    'rf__n_estimators': [5600, 5700, 5800],
    'rf__max_depth': [72,74,76],
    'rf__min_samples_leaf': [1, 2, 3],
    'rf__min_samples_split': [3, 4, 5],
    'rf__ccp_alpha': [0.01, 0.1, 1.0]
}

gs_pipe2 = GridSearchCV(pipe2,
                        param_grid=pipe2_params,
                        cv = 5,
                        n_jobs = -1)

gs_pipe2.fit(X_train, y_train)

print(f'Best Score: {gs_pipe2.best_score_}')
print(f'Best Params: {gs_pipe2.best_params_}')
print(f'Training Score: {gs_pipe2.score(X_train, y_train)}')
print(f'Testing Score: {gs_pipe2.score(X_test, y_test)}')

pipe3 = Pipeline([
    ('cvec', CountVectorizer(stop_words=stop_words_list)),
    ('log', LogisticRegression())
])

pipe3_params = {
    'log__penalty': ['l2'],
    'log__solver': ['newton-cg', 'sag', 'lbfgs'],
    'log__C': [0.07, 0.08, 0.09]
}

gs_pipe3 = GridSearchCV(pipe3,
                        param_grid=pipe3_params,
                        cv = 5,
                        n_jobs = -1)

gs_pipe3.fit(X_train, y_train)

print(f'Best Score: {gs_pipe3.best_score_}')
print(f'Best Params: {gs_pipe3.best_params_}')
print(f'Training Score: {gs_pipe3.score(X_train, y_train)}')
print(f'Testing Score: {gs_pipe3.score(X_test, y_test)}')

pipe3 = Pipeline([
    ('cvec', CountVectorizer(stop_words=stop_words_list)),
    ('log', LogisticRegression(max_iter = 10000))
])

pipe3_params = {
    'log__penalty': ['l1'],
    'log__solver': ['liblinear', 'saga'],
    'log__C': [0.9, 1, 1.1]
}

gs_pipe3 = GridSearchCV(pipe3,
                        param_grid=pipe3_params,
                        cv = 5,
                        n_jobs = -1)

gs_pipe3.fit(X_train, y_train)

print(f'Best Score: {gs_pipe3.best_score_}')
print(f'Best Params: {gs_pipe3.best_params_}')
print(f'Training Score: {gs_pipe3.score(X_train, y_train)}')
print(f'Testing Score: {gs_pipe3.score(X_test, y_test)}')

lvl1_est = [
    ('nb', MultinomialNB(alpha = 0.7)),
    ('rf', RandomForestClassifier(max_depth = 92, 
                                  min_samples_leaf = 1,
                                  min_samples_split = 5, 
                                  n_estimators = 50,
                                  ccp_alpha = 0.01)),
    ('log', LogisticRegression(penalty='l2', 
                               solver='sag', 
                               C = 0.08))
]

stacked_1 = StackingClassifier(estimators=lvl1_est, 
                               final_estimator=LogisticRegression(),
                               n_jobs = -1)

pipe_cvec_1 = Pipeline([
    ('cvec', CountVectorizer(stop_words=stop_words_list,
                            max_df=0.1,
                            min_df = 2,
                            max_features=4000,
                            ngram_range= (1, 1))),
    ('s1', stacked_1)
])

print(cross_val_score(pipe_cvec_1, X_train, y_train).mean())
pipe_cvec_1.fit(X_train, y_train)
print(f'Training Score: {pipe_cvec_1.score(X_train, y_train)}')
print(f'Testing Score: {pipe_cvec_1.score(X_test, y_test)}')

lvl1_est = [
    ('nb', MultinomialNB(alpha = 0.7)),
    ('rf', RandomForestClassifier(max_depth = 92, 
                                  min_samples_leaf = 1,
                                  min_samples_split = 5, 
                                  n_estimators = 50,
                                  ccp_alpha = 0.01)),
    ('log', LogisticRegression(penalty='l2', 
                               solver='sag', 
                               C = 0.08))
]

stacked_1 = StackingClassifier(estimators=lvl1_est, 
                               final_estimator=LogisticRegression(),
                               n_jobs = -1)

pipe_final = Pipeline([
    ('cvec', CountVectorizer(stop_words=stop_words_list,
                            max_df=0.1,
                            min_df = 2,
                            max_features=4000,
                            ngram_range= (1, 1))),
    ('s1', stacked_1)
])


pipe_final_params = {
    's1__final_estimator__penalty': ['l1', 'l2'],
    's1__final_estimator__solver': ['lbfgs', 'saga', 'sag', 'liblinear'],
    's1__final_estimator__C': [0.1, 1, 2]
}

gs_final = GridSearchCV(pipe_final,
                        param_grid= pipe_final_params,
                        cv = 5,
                        n_jobs = -1)
    
    
gs_final.fit(X_train, y_train)
print(f'Best Score: {gs_final.best_score_}')
print(f'Best Params: {gs_final.best_params_}')
print(f'Training Score: {gs_final.score(X_train, y_train)}')
print(f'Testing Score: {gs_final.score(X_test, y_test)}')
